[수정] step1 문자열 연산 operate 메서드를 Operator 클래스로 분리
